// src/Types/ClinicAdmin.ts
export default interface ClinicType {
  clinic_name?: any;
  role_id?: any;
  id?: string;
  first_name?: string;
  mid_name?: string;
  last_name?: string;
  contact_no?: string;
  role?: string;
  degree: string | undefined;
  clinic_id?: string[];
  clinic?: string;
  gender?: string;
  email?: string;
  password?: string;
  program_id?: string[];
}